Add an instance variable to allow shifting the column address an arbitrary amount
Found this lurking as a const in the [old C library][1] as (`ST7565_STARTBYTES`), and although the value presented there wasn't working for my display, I was able to get my display working correctly with a different value. Decided to make it an instance variable so it can be adjusted as needed (alongside the pagemap list, which I also needed to adjust for my display).

[1]: https://github.com/adafruit/ST7565-LCD/blob/d40d9d832be8a6ce26d6462e85de3dcf439942fc/ST7565/ST7565.cpp#L593